[Ref] Silence warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@22683 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/DialogBase.cpp
@@ -114,7 +114,7 @@ BOOL DialogBase::OnToolTipText(UINT, NMHDR *pNMHDR, LRESULT *pResult)
 	if(s.IsEmpty())
 		return FALSE;
 
-	if(s.GetLength() < std::size(pTTT->szText))
+	if(s.GetLength() < static_cast<int>(std::size(pTTT->szText)))
 	{
 		mpt::String::WriteCStringBuf(pTTT->szText) = s;
 	} else
